% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_accuracy.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\title{Calculate Accuracy}
\usage{
calculate_accuracy(true_val, pred_val, threshold = 0.5)
}
\arguments{
\item{true_val}{ground truth values in range {0, 1}}

\item{pred_val}{predicted probability in range [0, 1]}

\item{threshold}{probability threshold, can be a single value or a vector, default 0.5}
}
\value{
mean accuracy
}
\description{
Function to calculate accuracy based on ground truth and
predicted probability. First threshold is applied and then mean accuracy
across samples is calculated. Function also calculates number of true positive, false positive, true negative and false negative examples.
}
\examples{
 calculate_accuracy(true_val, pred_val) or calculate_accuracy(true_val, pred_val, threshold)
}
