% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_precsion_recall.R
\name{calculate_precsion_recall}
\alias{calculate_precsion_recall}
\title{Calculate precision and recall}
\usage{
calculate_precsion_recall(accuracy)
}
\arguments{
\item{accuracy}{metrics}
}
\value{
precision and recall in a matrix of size length(thresholds)x2
}
\description{
Function to calculate precision and recall based on metrics object
}
\examples{
calculate_precision_recall(accuracy)
}
