% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classifier}
\alias{classifier}
\title{Binary classification}
\usage{
classifier(training_x, training_y, testing)
}
\arguments{
\item{training_x}{input for model training}

\item{training_y}{expected output (true value)}

\item{testing}{input for model testing}
}
\value{
classification of testing vector
}
\description{
function for training model and classificing test data
}
\examples{
classifier(X, Y)
}
